+++ src/context.cc
@@ -29,11 +29,13 @@ Context::Context(std::shared_ptr<Config> config,
     std::shared_ptr<ClientManager> clients,
     std::shared_ptr<Mime> mime,
     std::shared_ptr<Database> database,
-    std::shared_ptr<Web::SessionManager> sessionManager)
+    std::shared_ptr<Web::SessionManager> sessionManager,
+    std::shared_ptr<ConverterManager> converterManager)
     : config(std::move(config))
     , clients(std::move(clients))
     , mime(std::move(mime))
     , database(std::move(database))
     , session_manager(std::move(sessionManager))
+    , converterManager(std::move(converterManager))
 {
 }