+++ src/context.h
@@ -35,6 +35,7 @@ class Database;
 class Mime;
 class Server;
 class UpdateManager;
+class ConverterManager;
 namespace Web {
 class SessionManager;
 } // namespace Web
@@ -45,7 +46,8 @@ public:
         std::shared_ptr<ClientManager> clients,
         std::shared_ptr<Mime> mime,
         std::shared_ptr<Database> database,
-        std::shared_ptr<Web::SessionManager> sessionManager);
+        std::shared_ptr<Web::SessionManager> sessionManager,
+        std::shared_ptr<ConverterManager> converterManager);
 
     std::shared_ptr<Config> getConfig() const
     {
@@ -72,12 +74,18 @@ public:
         return session_manager;
     }
 
+    std::shared_ptr<ConverterManager> getConverterManager() const
+    {
+        return converterManager;
+    }
+
 private:
     std::shared_ptr<Config> config;
     std::shared_ptr<ClientManager> clients;
     std::shared_ptr<Mime> mime;
     std::shared_ptr<Database> database;
     std::shared_ptr<Web::SessionManager> session_manager;
+    std::shared_ptr<ConverterManager> converterManager;
 };
 
 #endif // __CONTEXT_H__