+++ src/metadata/taglib_handler.h
@@ -70,7 +70,7 @@ private:
     static std::unique_ptr<TagLib::File> getOggFile(TagLib::IOStream& ioStream);
 
     void populateGenericTags(const std::shared_ptr<CdsItem>& item, const TagLib::File& file) const;
-    void populateAuxTags(const std::shared_ptr<CdsItem>& item, const TagLib::PropertyMap& propertyMap, const std::unique_ptr<StringConverter>& sc) const;
+    void populateAuxTags(const std::shared_ptr<CdsItem>& item, const TagLib::PropertyMap& propertyMap, const std::shared_ptr<StringConverter>& sc) const;
     static bool isValidArtworkContentType(std::string_view artMimetype);
     std::string getContentTypeFromByteVector(const TagLib::ByteVector& data) const;
     static void addArtworkResource(const std::shared_ptr<CdsItem>& item, const std::string& artMimetype);