+++ src/util/string_converter.h
@@ -34,10 +34,11 @@
 #ifndef __STRING_CONVERTER_H__
 #define __STRING_CONVERTER_H__
 
-#include <iconv.h>
-#include <memory>
-
 #include "config/config.h"
+
+#include <iconv.h>
+#include <map>
+#include <memory>
 
 class StringConverter {
 public:
@@ -56,34 +57,38 @@ public:
     std::string convert(std::string str, bool validate = false);
     bool validate(const std::string& str);
 
-    /// \brief internal (UTF-8) to filesystem
-    static std::unique_ptr<StringConverter> i2f(const std::shared_ptr<Config>& cm);
-
-    /// \brief filesystem to internal
-    static std::unique_ptr<StringConverter> f2i(const std::shared_ptr<Config>& cm);
-
-    /// \brief metadata to internal
-    static std::unique_ptr<StringConverter> m2i(ConfigVal option, const fs::path& location, const std::shared_ptr<Config>& cm);
-#ifdef HAVE_JS
-    /// \brief scripting to internal
-    static std::unique_ptr<StringConverter> j2i(const std::shared_ptr<Config>& cm);
-
-    /// \brief playlist to internal
-    static std::unique_ptr<StringConverter> p2i(const std::shared_ptr<Config>& cm);
-#endif
-#if defined(HAVE_JS) || defined(HAVE_TAGLIB) || defined(ATRAILERS) || defined(HAVE_MATROSKA)
-    /// \brief safeguard - internal to internal - needed to catch some
-    /// scenarious where the user may have forgotten to add proper conversion
-    /// in the script.
-    static std::unique_ptr<StringConverter> i2i([[maybe_unused]] const std::shared_ptr<Config>& cm);
-#endif
-
 protected:
     iconv_t cd;
     bool dirty {};
 
     std::string _convert(const std::string& str, bool validate,
         std::size_t* stoppedAt = nullptr);
+};
+
+class ConverterManager {
+public:
+    ConverterManager(const std::shared_ptr<Config>& cm);
+    virtual ~ConverterManager() = default;
+
+    const std::shared_ptr<StringConverter> m2i(ConfigVal option, const fs::path& location);
+    const std::shared_ptr<StringConverter> f2i() const;
+#if defined(HAVE_JS) || defined(HAVE_TAGLIB) || defined(ATRAILERS) || defined(HAVE_MATROSKA)
+    /// \brief safeguard - internal to internal - needed to catch some
+    /// scenarious where the user may have forgotten to add proper conversion
+    /// in the script.
+    const std::shared_ptr<StringConverter> i2i() const;
+#endif
+#ifdef HAVE_JS
+    /// \brief scripting to internal
+    const std::shared_ptr<StringConverter> j2i() const;
+    /// \brief playlist to internal
+    const std::shared_ptr<StringConverter> p2i() const;
+#endif
+
+protected:
+    std::shared_ptr<Config> config;
+    std::map<ConfigVal, std::string> charsets {};
+    std::map<std::string, std::shared_ptr<StringConverter>> converters {};
 };
 
 #endif // __STRING_CONVERTER_H__